Resolve "Add support for special holiday business hours"

See merge request cse1105/2019-2020/repos/oopp-group-28/repository-template!141

+++ server/src/main/java/nl/tudelft/oopp/demo/repositories/RestaurantHourRepository.java
@@ -9,9 +9,9 @@ import org.springframework.stereotype.Repository;
  */
 @Repository
 public interface RestaurantHourRepository extends JpaRepository<RestaurantHours, Integer> {
-    RestaurantHours findByRestaurant_IdAndDay(int restaurantId, int day);
+    RestaurantHours findByRestaurant_IdAndDay(int restaurantId, long day);
 
-    boolean existsByRestaurant_IdAndDay(int restaurantId, int day);
+    boolean existsByRestaurant_IdAndDay(int restaurantId, long day);
 
-    void deleteByRestaurant_IdAndDay(int restaurantId, int day);
+    void deleteByRestaurant_IdAndDay(int restaurantId, long day);
 }
